refactor: mv oapi schema builder to oapi package
