+++ pkg/checks/oapi/oapi.go
@@ -16,7 +16,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package checks
+package oapi
 
 import (
 	"github.com/caas-team/sparrow/pkg/checks/config"
