+++ pkg/checks/oapi/oapi_test.go
@@ -16,7 +16,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package checks
+package oapi
 
 import (
 	"reflect"
